Strip HTML tags from the summary
Summary is built from the first line of the description
(defined using Doc() in the RouteBuilder()).

Description may contain HTML, for emphasis in Swagger-UI
(or any other visualization tool), thus they should
be stripped from the one-line summary.

This is performed using a very basic regexp which removes
all "<[^>]*>" patterns.