Allow operations to return primitive types
Before this change, if an operation returned a primitive type, then
this would create a "ref" to it (which would never exist, because
it was a primitive type).

For example, if you returned a string, then you would get back a
"ref" of `#/definitions/string`, which is bad.  Now, the ref will
not be set, and the "type" will be set to "string" properly.