Add support for map types using "additional properties"
According to https://swagger.io/docs/specification/data-models/dictionaries/,
we can use a model's "additional properties" to create a functional
map structure.

This commit basically copies the array workflow and does something
similar for objects.  If the type of the map cannot be determined
(for example, if it is `map[string]interface{}`), then this will
just define an empty object with no additional properties, just
like it used to work prior to this change.