PostBuildSwaggerSchema handler for each model
Check if a type implements PostBuildSwaggerSchema interface and call PostBuildSwaggerSchemaHandler function with a schema created for that type.

Sometime there is a need in updating the schema of a model, the PostBuildSwaggerObjectHandler function also can be used, but it requries hardcoding the names of models, that i want to avoid.